Holiday Update
Continued to refine some visual aspects of InCA. Updated Readme with goals and to-do. Integrated more fitting types (polar) but broke parametric in the process. Isolated why a figure window opens with frame inspector; will continue to see if that behavior can be curtailed. Added tooltips to a majority of controls.